LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-pt-PT.html

<!DOCTYPE html><!-- webkit-test-runner [ language=pt-PT ] -->
<html>
<head>
<script src="../../../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "pt-pt");
</script>
<script src="../../../../../resources/js-test-post.js"></script>
</body>
</html>
